requirements.txt
Pillow>=9.3.0
matplotlib>=3.6.0
numpy>=1.23.4
pandas>=1.5.1
tqdm>=4.64.1